[ REFACTOR ] 인터뷰 navGraph 분리
