+++ app/src/main/java/com/lottomate/lottomate/presentation/screen/main/MainNavHost.kt
@@ -5,10 +5,10 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.compose.NavHost
 import com.lottomate.lottomate.data.error.LottoMateErrorType
 import com.lottomate.lottomate.presentation.screen.home.navigation.homeNavGraph
+import com.lottomate.lottomate.presentation.screen.interview.navigation.interviewNavGraph
 import com.lottomate.lottomate.presentation.screen.intro.navigation.introNavGraph
 import com.lottomate.lottomate.presentation.screen.login.navigation.loginNavGraph
 import com.lottomate.lottomate.presentation.screen.lounge.navigation.loungeNavGraph
-import com.lottomate.lottomate.presentation.screen.main.model.FullScreenType
 import com.lottomate.lottomate.presentation.screen.map.navigation.mapNavGraph
 import com.lottomate.lottomate.presentation.screen.pocket.navigation.pocketNavGraph
 import com.lottomate.lottomate.presentation.screen.setting.navigation.settingNavGraph
@@ -48,6 +48,12 @@ fun MainNavHost(
         )
 
         loungeNavGraph(
+            padding = padding,
+            navController = navigator.navController,
+            onShowErrorSnackBar = onShowErrorSnackBar,
+        )
+
+        interviewNavGraph(
             padding = padding,
             navController = navigator.navController,
             onShowErrorSnackBar = onShowErrorSnackBar,
